do wgml project more portable fix signed/unsigned char mismatch add font_number prototype and related input/output functions add macros for device/driver/font pointer manipulation remove cast as much as possible enable 64-bit nt build

+++ bld/wgml/c/wgmlmsg.c
@@ -109,7 +109,7 @@ int init_msgs( void )
 /*  get a msg text string                                                  */
 /***************************************************************************/
 
-int get_msg( msg_ids resid, char *buff, unsigned buff_len )
+int get_msg( msg_ids resid, char *buff, size_t buff_len )
 {
     if( WResLoadString( &Instance, resid + MsgShift, buff, buff_len ) != 0 ) {
         buff[0] = '\0';
